// src/app/modules/HiringPost/hiringPost.interface.ts
export interface THiring {
  title: string;
  hiringImage: string;
  jobNature: string;
  workingHours: string;
  workingDays: string;
  description: string;
  minSalary: number;
  maxSalary: number;
  currency: string;
  location: string;
  requirements: string;
  experience: string;
  benefits1: string;
  benefits2: string;
  benefits3: string;
  benefits4: string;
  benefits5: string;
  benefits6: string;
  benefits7: string;
  benefits8: string;
  benefits9: string;
  benefits10: string;
  companyName: string;
  responsibilities1: string;
  responsibilities2: string;
  responsibilities3: string;
  responsibilities4: string;
  responsibilities5: string;
  responsibilities6: string;
  InterviewRd1: string;
  InterviewRd2: string;
  InterviewRd3: string;
  applicationDeadline: string;
}